Move locale helper tests from feature to integration
